# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stellar Intel API",
    "version": "1.2.0",
    "description": "Intent router and anchor rate aggregation API for the Stellar Intel platform.\n\n## Public v1 surface\n\nThe stable, supported surface is namespaced under `/api/v1/...`; unversioned\nroutes are internal and may change without notice. Every v1 response follows\nthe hardening contract (see `lib/api/v1.ts`):\n\n- **Error envelope** — errors return `{ \"error\": { \"code\", \"message\", \"requestId\" } }`.\n- **Rate-limit headers** — `X-RateLimit-Limit`, `X-RateLimit-Remaining`,\n  `X-RateLimit-Reset` on every response; `Retry-After` on a `429`.\n- **Idempotency** — send an `Idempotency-Key` header on mutating endpoints\n  (e.g. `POST /api/v1/intent/offramp`); a retried request replays the original\n  response (`Idempotency-Replayed: true`) instead of creating a second intent.",
    "contact": {
      "name": "Stellar Intel",
      "url": "https://github.com/ezedike-evan/stellar-intel"
    }
  },
  "servers": [
    {
      "url": "https://stellar-intel.vercel.app/api/v1",
      "description": "Production (public v1)"
    },
    {
      "url": "https://stellar-intel.vercel.app",
      "description": "Production (internal, unversioned)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "schemas": {
...
      "OfframpRoute": {
        "type": "object",
        "properties": {
          "anchorId": {
            "type": "string"
          },
          "anchorDomain": {
            "type": "string"
          },
          "corridorId": {
            "type": "string"
          },
          "estimatedFee": {
            "type": "string"
          },
          "estimatedReceived": {
            "type": "string"
          }
        },
        "required": [
          "anchorId",
          "anchorDomain",
          "corridorId",
          "estimatedFee",
          "estimatedReceived"
        ]
      },
      "OfframpIntentResponse": {
        "type": "object",
        "properties": {
          "route": {
            "$ref": "#/components/schemas/OfframpRoute"
          },
          "unsignedTx": {
            "type": "string",
            "description": "XDR-encoded unsigned Stellar transaction"
          },
          "quoteId": {
            "type": "string",
            "description": "Hex-encoded SHA-256 quote identifier"
          }
        },
        "required": [
          "route",
          "unsignedTx",
          "quoteId"
        ]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error description"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "IntentRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "offramp"
            ]
          },
          "sourceAsset": {
            "type": "string",
            "minLength": 1
          },
          "destinationAsset": {
            "type": "string",
            "minLength": 1
          },
          "amount": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d+)?$"
          },
          "sender": {
            "type": "string",
            "minLength": 1,
            "description": "Stellar public key of the sender"
          },
          "recipient": {
            "type": "string",
            "minLength": 1,
            "description": "Destination address for the payout"
          }
        },
        "required": [
          "type",
          "sourceAsset",
          "destinationAsset",
          "amount",
          "sender",
          "recipient"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/intent/offramp": {
      "post": {
        "summary": "Submit an off-ramp intent",
        "description": "Resolves an anchor route for the given asset corridor, builds an unsigned Stellar payment transaction, and returns a quote ID.",
        "tags": [
          "Intent"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Route resolved and unsigned transaction built",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfframpIntentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or no route found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Transaction build failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}